requirements.txt
Cerberus==1.3.7
colorama==0.4.6
dnspython==2.7.0
greenlet==3.1.1
loguru==0.7.3
numpy==2.2.1
pandas==2.2.3
psycopg2==2.9.10
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.17.0
SQLAlchemy==2.0.37
typing_extensions==4.12.2
tzdata==2024.2
win32_setctime==1.2.0
